requirements.txt
followthemoney
pyicu
pyahocorasick
requests